Delay database connection until it is necessary.
Only creation and destruction of boards requires a database connection.
There is no need to connect to the database on other pages.

If database connection fails due to misconfiguration, and it is
attempted on the index or management page, quickinstall is rendered
useless as the user has no way of fixing the configuration.